+++ sql/types/number.go
@@ -1114,27 +1114,27 @@ func convertToUint64(t NumberTypeImpl_, v interface{}) (uint64, sql.ConvertInRan
 		return uint64(v.UTC().Unix()), sql.InRange, nil
 	case int:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.OutOfRange, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.OutOfRange, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int8:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.OutOfRange, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.OutOfRange, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int16:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.OutOfRange, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.OutOfRange, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int32:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.OutOfRange, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.OutOfRange, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int64:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.OutOfRange, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.OutOfRange, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case uint:
@@ -1204,35 +1204,35 @@ func convertToUint32(t NumberTypeImpl_, v interface{}) (uint32, sql.ConvertInRan
 	switch v := v.(type) {
 	case int:
 		if v < 0 {
-			return uint32(math.MaxUint32 - uint(-v-1)), sql.OutOfRange, nil
-		} else if v > math.MaxUint32 {
+			return uint32(math.MaxUint32 - uint64(-v-1)), sql.OutOfRange, nil
+		} else if int64(v) > int64(math.MaxUint32) {
 			return uint32(math.MaxUint32), sql.OutOfRange, nil
 		}
 		return uint32(v), sql.InRange, nil
 	case int8:
 		if v < 0 {
-			return uint32(math.MaxUint32 - uint(-v-1)), sql.OutOfRange, nil
-		} else if int(v) > math.MaxUint32 {
+			return uint32(math.MaxUint32 - uint64(-v-1)), sql.OutOfRange, nil
+		} else if int64(v) > int64(math.MaxUint32) {
 			return uint32(math.MaxUint32), sql.OutOfRange, nil
 		}
 		return uint32(v), sql.InRange, nil
 	case int16:
 		if v < 0 {
-			return uint32(math.MaxUint32 - uint(-v-1)), sql.OutOfRange, nil
-		} else if int(v) > math.MaxUint32 {
+			return uint32(math.MaxUint32 - uint64(-v-1)), sql.OutOfRange, nil
+		} else if int64(v) > int64(math.MaxUint32) {
 			return uint32(math.MaxUint32), sql.OutOfRange, nil
 		}
 		return uint32(v), sql.InRange, nil
 	case int32:
 		if v < 0 {
-			return uint32(math.MaxUint32 - uint(-v-1)), sql.OutOfRange, nil
-		} else if int(v) > math.MaxUint32 {
+			return uint32(math.MaxUint32 - uint64(-v-1)), sql.OutOfRange, nil
+		} else if int64(v) > int64(math.MaxUint32) {
 			return uint32(math.MaxUint32), sql.OutOfRange, nil
 		}
 		return uint32(v), sql.InRange, nil
 	case int64:
 		if v < 0 {
-			return uint32(math.MaxUint32 - uint(-v-1)), sql.OutOfRange, nil
+			return uint32(math.MaxUint32 - uint64(-v-1)), sql.OutOfRange, nil
 		} else if v > math.MaxUint32 {
 			return uint32(math.MaxUint32), sql.OutOfRange, nil
 		}
